Added Features:
I added the feature to automatically detect if the server has magic_quotes_gpc() on, and if so, to un-sanitize the request data. Also, upon fatal error of eval'd code, the PHP Evaluation function now prints the error and the rest of the page, rather than terminating right after the fatal error.